Adding contants, renaming files, wiring up shipping section with discrete slots for variants and non-variants
