+++ plugins/woocommerce-admin/client/products/fills/constants.ts
@@ -1,17 +1,32 @@
-export const PRICING_SECTION_BASIC_ID = 'pricing/basic';
-export const PRICING_SECTION_TAXES_ID = 'pricing/taxes';
 export const DETAILS_SECTION_ID = 'general/details';
-export const INVENTORY_SECTION_ID = 'inventory/inventory';
-export const INVENTORY_SECTION_ADVANCED_ID = 'inventory/advanced';
 export const IMAGES_SECTION_ID = 'general/images';
 export const ATTRIBUTES_SECTION_ID = 'general/attributes';
+export const PRICING_SECTION_BASIC_ID = 'pricing/basic';
+export const PRICING_SECTION_TAXES_ID = 'pricing/taxes';
+export const INVENTORY_SECTION_ID = 'inventory/inventory';
+export const INVENTORY_SECTION_ADVANCED_ID = 'inventory/advanced';
 export const SHIPPING_SECTION_BASIC_ID = 'shipping/shipping';
 export const SHIPPING_SECTION_DIMENSIONS_ID = 'shipping/dimensions';
+
+export const VARIANT_PRICING_SECTION_BASIC_ID = 'variant/pricing/basic';
+export const VARIANT_PRICING_SECTION_TAXES_ID = 'variant/pricing/taxes';
+export const VARIANT_INVENTORY_SECTION_ID = 'variant/inventory/inventory';
+export const VARIANT_INVENTORY_SECTION_ADVANCED_ID =
+	'variant/inventory/advanced';
+export const VARIANT_SHIPPING_SECTION_BASIC_ID = 'variant/shipping/shipping';
+export const VARIANT_SHIPPING_SECTION_DIMENSIONS_ID =
+	'variant/shipping/dimensions';
 
 export const TAB_INVENTORY_ID = 'tab/inventory';
 export const TAB_GENERAL_ID = 'tab/general';
 export const TAB_SHIPPING_ID = 'tab/shipping';
 export const TAB_PRICING_ID = 'tab/pricing';
+export const TAB_OPTIONS_ID = 'tab/options';
+
+export const VARIANT_TAB_INVENTORY_ID = 'variant/tab/inventory';
+export const VARIANT_TAB_GENERAL_ID = 'variant/tab/general';
+export const VARIANT_TAB_SHIPPING_ID = 'variant/tab/shipping';
+export const VARIANT_TAB_PRICING_ID = 'variant/tab/pricing';
 
 export const PLUGIN_ID = 'woocommerce';
 export const PRODUCT_DETAILS_SLUG = 'product-details';
