+++ plugins/woocommerce-admin/client/products/fills/index.ts
@@ -1,11 +1,10 @@
 /**
  * Internal dependencies
  */
+import './product-form-api-fills';
 import './product-form-fills';
-import './product-form-tab-fills';
-import './product-form-variation-tab-fills';
+import './product-form-variation-fills';
 
-export * from './shipping-section/shipping-section-fills';
 export * from './details-section/details-section-fills';
 export * from './images-section/images-section-fills';
 export * from './attributes-section/attributes-section-fills';
