+++ plugins/woocommerce-admin/client/products/fills/product-form-fills.tsx
@@ -1,52 +1,130 @@
 /**
  * External dependencies
  */
+import { __ } from '@wordpress/i18n';
 import { registerPlugin } from '@wordpress/plugins';
-import { useSelect, resolveSelect } from '@wordpress/data';
 import {
-	EXPERIMENTAL_PRODUCT_FORM_STORE_NAME,
-	WCDataSelector,
-} from '@woocommerce/data';
-import { registerCoreProductFields } from '@woocommerce/components';
-
-registerCoreProductFields();
+	__experimentalWooProductTabItem as WooProductTabItem,
+	__experimentalWooProductSectionItem as WooProductSectionItem,
+	useFormContext,
+} from '@woocommerce/components';
+import { Product } from '@woocommerce/data';
+import { useMemo } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { Fields } from './product-form-field-fills';
-import { Sections } from './product-form-section-fills';
+import { ShippingSectionFills } from './shipping-section';
+import { OptionsSection } from '../sections/options-section';
+import { ProductVariationsSection } from '../sections/product-variations-section';
+import {
+	TAB_GENERAL_ID,
+	TAB_SHIPPING_ID,
+	TAB_INVENTORY_ID,
+	TAB_PRICING_ID,
+	TAB_OPTIONS_ID,
+	SHIPPING_SECTION_BASIC_ID,
+	SHIPPING_SECTION_DIMENSIONS_ID,
+	PLUGIN_ID,
+} from './constants';
 
-const Form = () => {
-	const { formData } = useSelect( ( select: WCDataSelector ) => {
-		return {
-			formData: select(
-				EXPERIMENTAL_PRODUCT_FORM_STORE_NAME
-			).getProductForm(),
-		};
-	} );
-
+const Tabs = () => {
+	const { values: product } = useFormContext< Product >();
+	const tabPropData = useMemo(
+		() => ( {
+			general: {
+				name: 'general',
+				title: __( 'General', 'woocommerce' ),
+			},
+			pricing: {
+				name: 'pricing',
+				title: __( 'Pricing', 'woocommerce' ),
+				disabled: !! product?.variations?.length,
+			},
+			inventory: {
+				name: 'inventory',
+				title: __( 'Inventory', 'woocommerce' ),
+				disabled: !! product?.variations?.length,
+			},
+			shipping: {
+				name: 'shipping',
+				title: __( 'Shipping', 'woocommerce' ),
+				disabled: !! product?.variations?.length,
+			},
+			options: {
+				name: 'options',
+				title: __( 'Options', 'woocommerce' ),
+			},
+		} ),
+		[ product.variations ]
+	);
 	return (
 		<>
-			{ formData && (
-				<>
-					<Sections sections={ formData.sections } />
-					<Fields fields={ formData.fields } />
-				</>
-			) }
+			<WooProductTabItem
+				id={ TAB_GENERAL_ID }
+				templates={ [ { name: 'tab/general', order: 1 } ] }
+				pluginId={ PLUGIN_ID }
+				tabProps={ tabPropData.general }
+			>
+				<WooProductSectionItem.Slot tab={ TAB_GENERAL_ID } />
+			</WooProductTabItem>
+			<WooProductTabItem
+				id={ TAB_PRICING_ID }
+				templates={ [ { name: 'tab/general', order: 3 } ] }
+				pluginId={ PLUGIN_ID }
+				tabProps={ tabPropData.pricing }
+			>
+				<WooProductSectionItem.Slot tab={ TAB_PRICING_ID } />
+			</WooProductTabItem>
+			<WooProductTabItem
+				id={ TAB_INVENTORY_ID }
+				templates={ [ { name: 'tab/general', order: 5 } ] }
+				pluginId={ PLUGIN_ID }
+				tabProps={ tabPropData.inventory }
+			>
+				<WooProductSectionItem.Slot tab={ TAB_INVENTORY_ID } />
+			</WooProductTabItem>
+			<WooProductTabItem
+				id={ TAB_SHIPPING_ID }
+				templates={ [ { name: 'tab/general', order: 7 } ] }
+				pluginId={ PLUGIN_ID }
+				tabProps={ tabPropData.shipping }
+			>
+				<WooProductSectionItem.Slot
+					tab={ TAB_SHIPPING_ID }
+					fillProps={ { product } }
+				/>
+			</WooProductTabItem>
+			{ window.wcAdminFeatures[ 'product-variation-management' ] ? (
+				<WooProductTabItem
+					id={ TAB_OPTIONS_ID }
+					templates={ [ { name: 'tab/general', order: 9 } ] }
+					pluginId={ PLUGIN_ID }
+					tabProps={ tabPropData.options }
+				>
+					<>
+						<OptionsSection />
+						<ProductVariationsSection />
+					</>
+				</WooProductTabItem>
+			) : null }
 		</>
 	);
 };
 
-/**
- * Preloading product form data, as product pages are waiting on this to be resolved.
- * The above Form component won't get rendered until the getProductForm is resolved.
- */
-resolveSelect( EXPERIMENTAL_PRODUCT_FORM_STORE_NAME ).getProductForm();
 registerPlugin( 'wc-admin-product-editor-form-fills', {
 	// @ts-expect-error 'scope' does exist. @types/wordpress__plugins is outdated.
 	scope: 'woocommerce-product-editor',
 	render: () => {
-		return <Form />;
+		return (
+			<>
+				<Tabs />
+				<ShippingSectionFills
+					tabId={ TAB_SHIPPING_ID }
+					basicSectionId={ SHIPPING_SECTION_BASIC_ID }
+					dimensionsSectionId={ SHIPPING_SECTION_DIMENSIONS_ID }
+				/>
+			</>
+		);
 	},
 } );
