+++ plugins/woocommerce-admin/client/products/fills/shipping-section/index.ts
@@ -4,3 +4,4 @@ export * from './shipping-field-dimensions-length';
 export * from './shipping-field-dimensions-height';
 export * from './shipping-field-dimensions-weight';
 export * from './types';
+export * from './shipping-section-fills';
