+++ plugins/woocommerce-admin/client/products/fills/shipping-section/shipping-section-fills.tsx
@@ -7,7 +7,6 @@ import {
 	__experimentalWooProductFieldItem as WooProductFieldItem,
 	__experimentalProductSectionLayout as ProductSectionLayout,
 } from '@woocommerce/components';
-import { registerPlugin } from '@wordpress/plugins';
 import { PartialProduct, OPTIONS_STORE_NAME } from '@woocommerce/data';
 import { useSelect } from '@wordpress/data';
 import { useState } from '@wordpress/element';
@@ -26,12 +25,7 @@ import {
 	DimensionPropsType,
 	ShippingDimensionsPropsType,
 } from './index';
-import {
-	PLUGIN_ID,
-	SHIPPING_SECTION_BASIC_ID,
-	SHIPPING_SECTION_DIMENSIONS_ID,
-	TAB_SHIPPING_ID,
-} from '../constants';
+import { PLUGIN_ID } from '../constants';
 import {
 	ShippingDimensionsImage,
 	ShippingDimensionsImageProps,
@@ -40,7 +34,17 @@ import { useProductHelper } from '../../use-product-helper';
 
 import './shipping-section.scss';
 
-const ShippingSection = () => {
+type ShippingSectionProps = {
+	basicSectionId: string;
+	dimensionsSectionId: string;
+	tabId: string;
+};
+
+export const ShippingSectionFills: React.FC< ShippingSectionProps > = ( {
+	basicSectionId,
+	dimensionsSectionId,
+	tabId,
+} ) => {
 	const [ highlightSide, setHighlightSide ] =
 		useState< ShippingDimensionsImageProps[ 'highlight' ] >();
 	const { parseNumber } = useProductHelper();
@@ -73,8 +77,8 @@ const ShippingSection = () => {
 	return (
 		<>
 			<WooProductSectionItem
-				id={ SHIPPING_SECTION_BASIC_ID }
-				tabs={ [ { name: TAB_SHIPPING_ID, order: 1 } ] }
+				id={ basicSectionId }
+				tabs={ [ { name: tabId, order: 1 } ] }
 				pluginId={ PLUGIN_ID }
 			>
 				<ProductSectionLayout
@@ -87,7 +91,7 @@ const ShippingSection = () => {
 					<Card>
 						<CardBody className="product-shipping-section__classes">
 							<WooProductFieldItem.Slot
-								section={ SHIPPING_SECTION_BASIC_ID }
+								section={ basicSectionId }
 							/>
 						</CardBody>
 					</Card>
@@ -104,9 +108,7 @@ const ShippingSection = () => {
 								<div className="product-shipping-section__dimensions-body-col">
 									{ hasResolvedUnits && (
 										<WooProductFieldItem.Slot
-											section={
-												SHIPPING_SECTION_DIMENSIONS_ID
-											}
+											section={ dimensionsSectionId }
 											fillProps={
 												{
 													setHighlightSide,
@@ -129,7 +131,7 @@ const ShippingSection = () => {
 			</WooProductSectionItem>
 			<WooProductFieldItem
 				id="shipping/class"
-				sections={ [ { name: SHIPPING_SECTION_BASIC_ID, order: 1 } ] }
+				sections={ [ { name: basicSectionId, order: 1 } ] }
 				pluginId={ PLUGIN_ID }
 			>
 				{ ( { product }: ProductShippingSectionPropsType ) => (
@@ -138,9 +140,7 @@ const ShippingSection = () => {
 			</WooProductFieldItem>
 			<WooProductFieldItem
 				id="shipping/dimensions/width"
-				sections={ [
-					{ name: SHIPPING_SECTION_DIMENSIONS_ID, order: 1 },
-				] }
+				sections={ [ { name: dimensionsSectionId, order: 1 } ] }
 				pluginId={ PLUGIN_ID }
 			>
 				{ ( { ...props }: ShippingDimensionsPropsType ) => (
@@ -149,9 +149,7 @@ const ShippingSection = () => {
 			</WooProductFieldItem>
 			<WooProductFieldItem
 				id="shipping/dimensions/length"
-				sections={ [
-					{ name: SHIPPING_SECTION_DIMENSIONS_ID, order: 3 },
-				] }
+				sections={ [ { name: dimensionsSectionId, order: 3 } ] }
 				pluginId={ PLUGIN_ID }
 			>
 				{ ( { ...props }: ShippingDimensionsPropsType ) => (
@@ -160,9 +158,7 @@ const ShippingSection = () => {
 			</WooProductFieldItem>
 			<WooProductFieldItem
 				id="shipping/dimensions/height"
-				sections={ [
-					{ name: SHIPPING_SECTION_DIMENSIONS_ID, order: 5 },
-				] }
+				sections={ [ { name: dimensionsSectionId, order: 5 } ] }
 				pluginId={ PLUGIN_ID }
 			>
 				{ ( { ...props }: ShippingDimensionsPropsType ) => (
@@ -171,9 +167,7 @@ const ShippingSection = () => {
 			</WooProductFieldItem>
 			<WooProductFieldItem
 				id="shipping/dimensions/weight"
-				sections={ [
-					{ name: SHIPPING_SECTION_DIMENSIONS_ID, order: 7 },
-				] }
+				sections={ [ { name: dimensionsSectionId, order: 7 } ] }
 				pluginId={ PLUGIN_ID }
 			>
 				<ShippingDimensionsWeightField />
@@ -181,9 +175,3 @@ const ShippingSection = () => {
 		</>
 	);
 };
-
-registerPlugin( 'wc-admin-product-editor-shipping-section', {
-	// @ts-expect-error 'scope' does exist. @types/wordpress__plugins is outdated.
-	scope: 'woocommerce-product-editor',
-	render: () => <ShippingSection />,
-} );
